Move AddBookBtn.js and RemoveBookBtn.js components to scr/components
